Shrunk block_size_t to a uint16_t.
This has the effect of reducing sizeof(gc_entry_t::block_info_t) by 4 bytes.

The value is already stored as a uint16_t in the LBA.

+++ src/buffer_cache/alt.cc
@@ -808,7 +808,7 @@ buf_read_t::~buf_read_t() {
     lock_->access_ref_count_--;
 }
 
-const void *buf_read_t::get_data_read(uint32_t *block_size_out) {
+const void *buf_read_t::get_data_read(uint16_t *block_size_out) {
     page_t *page = lock_->get_held_page_for_read();
     if (!page_acq_.has()) {
         page_acq_.init(page, &lock_->cache()->page_cache_,
@@ -830,7 +830,7 @@ buf_write_t::~buf_write_t() {
     lock_->access_ref_count_--;
 }
 
-void *buf_write_t::get_data_write(uint32_t block_size) {
+void *buf_write_t::get_data_write(uint16_t block_size) {
     page_t *page = lock_->get_held_page_for_write();
     if (!page_acq_.has()) {
         page_acq_.init(page, &lock_->cache()->page_cache_,
